+++ src/Vortice.Direct2D1/DirectWrite/DWrite.cs
@@ -12,6 +12,18 @@ namespace Vortice.DirectWrite
         ///  Try to create new instance of <see cref="IDWriteFactory"/>.
         /// </summary>
         /// <typeparam name="T">Type based on <see cref="IDWriteFactory"/>.</typeparam>
+        /// <param name="factoryType">The <see cref="IDWriteFactory"/> type.</param>
+        /// <returns>Return the <see cref="Result"/>.</returns>
+        public static T DWriteCreateFactory<T>(FactoryType factoryType = FactoryType.Shared) where T : IDWriteFactory
+        {
+            DWriteCreateFactory(factoryType, typeof(IDWriteFactory).GUID, out IntPtr nativePtr).CheckError();
+            return MarshallingHelpers.FromPointer<T>(nativePtr);
+        }
+
+        /// <summary>
+        ///  Try to create new instance of <see cref="IDWriteFactory"/>.
+        /// </summary>
+        /// <typeparam name="T">Type based on <see cref="IDWriteFactory"/>.</typeparam>
         /// <param name="factory">The <see cref="IDWriteFactory"/> being created.</param>
         /// <returns>Return the <see cref="Result"/>.</returns>
         public static Result DWriteCreateFactory<T>(out T? factory) where T : IDWriteFactory